refactor(tests): dissolve recon test grab-bag into per-source mirrors
tests/tools/recon/test_recon_tools.py (672) was the last recon test
grab-bag - a mix of scope-guard, run_recon, and per-binary smoke tests
that predated the per-source test layout the rest of the dir now follows.
Split each class to the module it exercises, no coverage change (53
tests redistributed, none dropped or duplicated):

  test_scope.py             <- TestHostOf / TestFilterInScope /
                               TestInScopeTypedAliases (scope.py)
  test_recon.py             <- TestSeedingConstants / TestRunRecon
                               (__init__.py: constants + orchestrator)
  test_cert_transparency.py <- TestCertTransparency
  test_waybackurls.py       <- TestHistoricalUrls
  test_dirfuzz.py           <- TestDiscoverPaths
  test_subfinder.py         += TestEnumerateSubdomainsParsing (parse /
                               dedup / missing-binary; complements the
                               existing flag tests)
  httpx/test_scanner.py     -> probe_endpoints parse/stdin/missing-binary
                               cases fold into a single TestProbeEndpoints
  nmap/test_scanner.py      += TestPortScan (the {host: ports} flatten had
                               no package-level coverage)

Largest resulting file is test_scope.py at 198 lines, well under the
450 bar. The subfinder parsing tests adopt the file's existing
_completed() helper and _run patch target for consistency.
